changed repos to the repositories folder, refactored repo name in services
