+++ src/main/java/com/swm/datatracker/controllers/WorkOrderController.java
@@ -1,19 +1,17 @@
 package com.swm.datatracker.controllers;
 
-import com.swm.datatracker.services.WorkOrderRepo;
+import com.swm.datatracker.respositories.WorkOrderRepository;
 import com.swm.datatracker.services.WorkOrderService;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class WorkOrderController {
-    private WorkOrderRepo workOrderRepo;
+    private WorkOrderRepository workOrderRepository;
     private WorkOrderService workOrderService;
 
-    public WorkOrderController(WorkOrderService service, WorkOrderRepo workOrderRepo){
-        this.workOrderRepo = workOrderRepo;
+    public WorkOrderController(WorkOrderService service, WorkOrderRepository workOrderRepository){
+        this.workOrderRepository = workOrderRepository;
         this.workOrderService = service;
     }
