+++ src/main/java/com/swm/datatracker/respositories/RolesRepository.java
@@ -1,4 +1,4 @@
-package com.swm.datatracker.services;
+package com.swm.datatracker.respositories;
 
 import com.swm.datatracker.models.UserRole;
 import org.springframework.data.jpa.repository.Query;
