+++ src/main/java/com/swm/datatracker/respositories/WorkOrderRepository.java
@@ -1,10 +1,10 @@
-package com.swm.datatracker.services;
+package com.swm.datatracker.respositories;
 
 import com.swm.datatracker.models.WorkOrder;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkOrderRepo extends CrudRepository<WorkOrder, Long> {
+public interface WorkOrderRepository extends CrudRepository<WorkOrder, Long> {
 
 }
